Create neophyte
I try to use the search algorithm to find the way to cut the cake with one raisin..I cannot get the correct piece in the original setting, but it still produces some correct way of cutting the cake. I started from the leftmost and upmost raisin, and try to cut the piece with the right area. I attach the vertically longest piece at the beginning and try to finish one horizontal coordinate at a specific row with other raisins. And then attach the horizontally longest possible piece lastly..